acquisition/pyproject.toml
[project]
name = "support-copilot-acquisition"
version = "0.1.0"
description = "Acquisition pipeline: ingest, chunk, embed, store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "google-cloud-aiplatform>=1.74.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.47.0",
    "langchain-text-splitters>=0.3.0",
    "lxml>=5.3.0",
    "numpy>=1.26.0",
    "openai>=1.58.0",
    "pymupdf>=1.25.0",
    "scikit-learn>=1.6.0",
    "tiktoken>=0.8.0",
    "typer>=0.15.0",
]
